실습

+++ class/221018/articles/views.py
@@ -3,6 +3,8 @@ from articles.forms import ArticleForm, CommentForm
 
 from articles.models import Article, Comment
 
+from django.contrib.auth.decorators import login_required
+
 # Create your views here.
 
 
@@ -14,11 +16,14 @@ def index(request):
     return render(request, "articles/index.html", context)
 
 
+@login_required
 def create(request):
     if request.method == "POST":
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect("articles:detail", article.pk)
     else:
         form = ArticleForm()
@@ -40,38 +45,54 @@ def detail(request, pk):
     return render(request, "articles/detail.html", context)
 
 
+@login_required
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    if request.method == "POST":
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect("articles:detail", article.pk)
+    if request.user == article.user:
+        if request.method == "POST":
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect("articles:detail", article.pk)
+        else:
+            form = ArticleForm(instance=article)
+        context = {
+            "form": form,
+            "article": article,
+        }
+        return render(request, "articles/update.html", context)
     else:
-        form = ArticleForm(instance=article)
-    context = {
-        "form": form,
-        "article": article,
-    }
-    return render(request, "articles/update.html", context)
+        from django.http import HttpResponseForbidden
+
+        return HttpResponseForbidden()
 
 
 def delete(request, pk):
-    Article.objects.get(pk=pk).delete()
-    return redirect("articles:index")
+    article = Article.objects.get(pk=pk)
+    if request.user == article.user:
+        article.delete()
+        return redirect("articles:index")
+    else:
+        from django.http import HttpResponseForbidden
+
+        return HttpResponseForbidden()
 
 
+@login_required
 def comments_create(request, pk):
     article = Article.objects.get(pk=pk)
     comment_form = CommentForm(request.POST)
     if comment_form.is_valid():
         comment = comment_form.save(commit=False)
         comment.article = article
+        comment.user = request.user
         comment.save()
     return redirect("articles:detail", article.pk)
 
 
+@login_required
 def comments_delete(request, article_pk, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
+    if request.user == comment.user:
+        comment.delete()
     return redirect("articles:detail", article_pk)
